requirements.txt
# Core app framework
streamlit>=1.24.0
streamlit-option-menu>=0.3.2

# Data processing
numpy>=1.21.0
pandas>=1.3.0

# Machine Learning
scikit-learn>=1.0.0,<1.4.0
xgboost>=1.6.0

# API + environment management
requests>=2.25.0
python-dotenv>=0.19.0

# Visualization
plotly>=5.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Utilities
geopy>=2.2.0
joblib>=1.1.0

# PDF generation (optional)
fpdf2>=2.7.0
